NF: Split RecursivePictureMenuTest
With part that requires Android and part that don't
